+++ backends/hip-gen/ceed-hip-gen.c
@@ -1,0 +1,59 @@
+// Copyright (c) 2017-2018, Lawrence Livermore National Security, LLC.
+// Produced at the Lawrence Livermore National Laboratory. LLNL-CODE-734707.
+// All Rights reserved. See files LICENSE and NOTICE for details.
+//
+// This file is part of CEED, a collection of benchmarks, miniapps, software
+// libraries and APIs for efficient high-order finite element and spectral
+// element discretizations for exascale applications. For more information and
+// source code availability see http://github.com/ceed.
+//
+// The CEED research is supported by the Exascale Computing Project 17-SC-20-SC,
+// a collaborative effort of two U.S. Department of Energy organizations (Office
+// of Science and the National Nuclear Security Administration) responsible for
+// the planning and preparation of a capable exascale ecosystem, including
+// software, applications, hardware, advanced system engineering and early
+// testbed platforms, in support of the nation's exascale computing imperative.
+
+#include <string.h>
+#include <stdarg.h>
+#include "ceed-hip-gen.h"
+
+//------------------------------------------------------------------------------
+// Backend init
+//------------------------------------------------------------------------------
+static int CeedInit_Hip_gen(const char *resource, Ceed ceed) {
+  int ierr;
+  const int nrc = 8; // number of characters in resource
+  if (strncmp(resource, "/gpu/hip/gen", nrc))
+    // LCOV_EXCL_START
+    return CeedError(ceed, 1, "Hip backend cannot use resource: %s", resource);
+  // LCOV_EXCL_STOP
+
+  Ceed ceedshared;
+  CeedInit("/gpu/hip/shared", &ceedshared);
+  ierr = CeedSetDelegate(ceed, ceedshared); CeedChk(ierr);
+
+  Ceed_Hip_gen *data;
+  ierr = CeedCalloc(1, &data); CeedChk(ierr);
+  ierr = CeedSetData(ceed, data); CeedChk(ierr);
+  ierr = CeedHipInit(ceed, resource, nrc); CeedChk(ierr);
+
+  const char fallbackresource[] = "/gpu/hip/ref";
+  ierr = CeedSetOperatorFallbackResource(ceed, fallbackresource); CeedChk(ierr);
+
+  ierr = CeedSetBackendFunction(ceed, "Ceed", ceed, "QFunctionCreate",
+                                CeedQFunctionCreate_Hip_gen); CeedChk(ierr);
+  ierr = CeedSetBackendFunction(ceed, "Ceed", ceed, "OperatorCreate",
+                                CeedOperatorCreate_Hip_gen); CeedChk(ierr);
+  ierr = CeedSetBackendFunction(ceed, "Ceed", ceed, "Destroy",
+                                CeedDestroy_Hip); CeedChk(ierr);
+  return 0;
+}
+
+//------------------------------------------------------------------------------
+// Register backend
+//------------------------------------------------------------------------------
+CEED_INTERN int CeedRegister_Hip_Gen(void) {
+  return CeedRegister("/gpu/hip/gen", CeedInit_Hip_gen, 20);
+}
+//------------------------------------------------------------------------------